Revert "code clean up"
This reverts commit 3b55dd26afee9b8c4ae0db7a8a09ce68eb33efb2.

+++ VncSharp/Encodings/PixelReader32.cs
@@ -15,6 +15,7 @@
 // along with this program; if not, write to the Free Software
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
+using System;
 using System.IO;
 
 namespace VncSharp.Encodings
@@ -33,10 +34,10 @@ namespace VncSharp.Encodings
 			// Read the pixel value
 			byte[] b = reader.ReadBytes(4);
 
-            uint pixel = ((uint)b[0]) & 0xFF |
-                                ((uint)b[1]) << 8 |
-                                ((uint)b[2]) << 16 |
-                                ((uint)b[3]) << 24;
+            uint pixel = (uint)(((uint)b[0]) & 0xFF | 
+                                ((uint)b[1]) << 8   | 
+                                ((uint)b[2]) << 16  | 
+                                ((uint)b[3]) << 24);
 
 			// Extract RGB intensities from pixel
 			byte red   = (byte) ((pixel >> framebuffer.RedShift)   & framebuffer.RedMax);
