+++ VncSharp/zlib.NET/SupportClass.cs
@@ -1,4 +1,7 @@
 
+using System;
+
+
 namespace ComponentAce.Compression.Libs.zlib
 {
 	public class SupportClass
@@ -115,7 +118,7 @@ namespace ComponentAce.Compression.Libs.zlib
 				return -1;
                 
 			for(int i = start; i < start + bytesRead; i++)
-				target[i] = receiver[i];
+				target[i] = (byte)receiver[i];
                 
 			return bytesRead;
 		}
